requirements.txt
langchain
openai
streamlit
pandas
duckduckgo-search
python-dotenv
pdfplumber
python-docx